pitcher_program_app/requirements.txt
python-telegram-bot>=20.0
anthropic>=0.39.0
httpx>=0.27.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.32.0
